# docs/openapi.yaml
openapi: 3.0.0
info:
  title: AI Travel Planner API
  description: API for AI-powered travel planning application
  version: 1.0.0
  contact:
    name: API Support
    email: support@ai-travel-planner.com
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.ai-travel-planner.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Place:
      type: object
      properties:
        place_id:
          type: string
        name:
          type: string
        category:
          type: array
          items:
            type: string
        location:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        rating:
          type: number
        opening_hours:
          type: object
        entry_fee:
          type: number
        city:
          type: string
        country:
          type: string
        tags:
          type: array
          items:
            type: string
        last_updated:
          type: string
          format: date-time
    Itinerary:
      type: object
      properties:
        _id:
          type: string
        user_id:
          type: string
        destination:
          type: string
        days:
          type: number
        budget:
          type: number
        plan:
          type: array
          items:
            type: object
            properties:
              day:
                type: number
              places:
                type: array
                items:
                  type: object
                  properties:
                    place_id:
                      type: string
                    name:
                      type: string
                    start_time:
                      type: string
                    end_time:
                      type: string
        generated_on:
          type: string
          format: date-time
        format:
          type: string
          enum: [standard, detailed, compact]
...
paths:
...
  /places/search:
    get:
      summary: Search for places
      tags: [Places]
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search query
        - in: query
          name: city
          schema:
            type: string
          description: Filter by city
        - in: query
          name: lat
          schema:
            type: number
          description: Latitude for location-based search
        - in: query
          name: lng
          schema:
            type: number
          description: Longitude for location-based search
      responses:
        '200':
          description: List of places
          content:
            application/json:
              schema:
                type: object
                properties:
                  places:
                    type: array
                    items:
                      $ref: '#/components/schemas/Place'
                  total:
                    type: number
                  has_more:
                    type: boolean
  /itineraries:
    post:
      summary: Generate and save a new itinerary
      tags: [Itineraries]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - destination
                - days
                - budget
              properties:
                destination:
                  type: string
                days:
                  type: number
                budget:
                  type: number
                preferences:
                  type: object
      responses:
        '201':
          description: Itinerary created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  itinerary:
                    $ref: '#/components/schemas/Itinerary'
        '400':
          description: Validation error
        '503':
          description: ML service unavailable

    get:
      summary: Get user's itineraries
      tags: [Itineraries]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: number
          description: Number of results
        - in: query
          name: offset
          schema:
            type: number
          description: Offset for pagination
      responses:
        '200':
          description: List of itineraries
  /memories:
    post:
      summary: Create a new memory with image upload
      tags: [Memories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - image
              properties:
                title:
                  type: string
                image:
                  type: string
                  format: binary
                note:
                  type: string
      responses:
        '201':
          description: Memory created successfully
...